Install script now works when no dotfiles exist

+++ install.sh
@@ -7,11 +7,6 @@ dir=~/dotfiles # dotfiles directory
 olddir=~/dotfiles_old # old dotfiles backup directory
 files="gitconfig gitignore profile zprofile zshrc" # list of files/folders to symlink in homedir
 
-# create dotfiles_old in homedir
-echo "Creating $olddir to backup any existing dotfiles in ~/..."
-mkdir -p $olddir
-echo "...done."
-
 # change to the dotfiles directory
 echo "Changing to the $dir directory..." 
 cd $dir
@@ -20,7 +15,11 @@ echo "...done."
 # move any existing dotfiles in homedir to dotfiles_old directory
 # create symlinks from the homedir to the dotfiles directory
 for file in $files; do
-	mv ~/.$file ~/dotfiles_old/
+	if [ -f ~/.$file ]
+	then
+		mkdir -p $olddir
+		mv ~/.$file ~/dotfiles_old/
+	fi
 	echo "Creating symlink to $file in home directory..."
 	ln -s $dir/$file ~/.$file
 done
